Create modules
I introduced modules in to the project. I divided the class Leaderboard into three separate modules.
